cv-service/requirements.txt
fastapi
uvicorn[standard]
insightface
onnxruntime
faiss-cpu
numpy
opencv-python
pydantic
python-multipart
